pointerevents/pointerevent_element_click_fires_pointerevent.html

<!DOCTYPE HTML>
<title>element.click() fires a PointerEvent</title>
<link rel="help" href="https://github.com/w3c/pointerevents/pull/317">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<input id="target" style="margin: 20px">

<script>
  'use strict';
  let clickTest = async_test("element.click() fires a PointerEvent");
  let target = document.getElementById("target");

  target.addEventListener("click", clickTest.step_func((e)=>{
    assert_true(e instanceof PointerEvent, "click should be a PointerEvent");
  }));
  let eventWatcher = new EventWatcher(clickTest, target, ["click"]);
  eventWatcher.wait_for("click").then(()=>clickTest.done());
  target.click();
</script>
